requirements.txt
torch
transformers
numpy
Pillow